requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pandas
numpy
scikit-learn
joblib
